review laravel

+++ app/Http/Controllers/CategoryController.php
@@ -6,8 +6,7 @@ use Illuminate\Http\Request;
 
 class CategoryController extends Controller
 {
-	public function index()
-	{
+	public function index(){
 		$data['data'] = \DB::table('category')->get();
 		return view('category', $data);
 	}
@@ -24,7 +23,16 @@ class CategoryController extends Controller
 	}
 
 	public function delCat(Request $request){
-		\DB::table('category')->where('CatId', '=', $request->id)->delete();		
+		
+		$query = \DB::table('product')
+		->where('CatId', $request->id)
+		->get();
+
+		if (!empty ( $query )) {
+			return view('check',$query);
+		}	
+		\DB::table('category')->where('CatId', '=', $request->id)->delete();
+
 		return redirect('/category');
 	}
